s6rc.oeclass: automatically declare flags
There's a lot of boilerplate involved in creating an s6 service. On at
least two occations, I've missed declaring the *_s6rc_* as a recipe
flag, so the service/bundle got created without proper
dependencies/contents. Make this a little less error-prone and allow
omitting the "RECIPE_FLAGS += " lines by automatically defining them (as
class flags) based on the services. I'm not sure it matters, but avoid
adding duplicate declarations if they are already present in either
RECIPE_FLAGS (e.g. for all existing recipes) or CLASS_FLAGS.

This will of course not work for service names that are added to
S6RC_*_SERVICES depending on a USE flag setting, but there are none of
those currently in meta/base or meta/core, and it is still possible to
manually add to recipe flags.