image_ldconfig.oeclass: class for creating /etc/ld.so.cache
Having a valid /etc/ld.so.cache reduces the startup cost of running a
program. This is particularly noticable in exec-heavy workloads, such
as anything s6/execline based. On a concrete platform I'm working on,
the init phase drops from 6.8s to 6.3s by adding the class to the
rootfs recipe.