classes/qemu-cpu.oeclass: new class
We'll need a way to translate our notion of TARGET_CPU to the string
used by qemu, e.g. powerpc => ppc. For arm, armeb and aarch64, we
already agree. If and when the qemu recipe grows more use flags for
enabling other emulator targets, this should also get updated
accordingly.