class/elfwrapper: Set argv[0] to wrapper instead of wrapped cmd
When calling the wrapped command, set argv[0] to the wrapper
argv[0] instead of the full path to the wrapped command.

This should be safer, as argv[0] can be used by the wrapped
command (where as the full path to the wrapped command is not
safe to call directly).  Also, commands using argv[0] for usage
information looks nicer.

Signed-off-by: Esben Haabendal <esben@haabendal.dk>