class/manifest-version: Only consider matching tags
This adds USE_manifest_version_tag_match flag, which is used to filter tags
used to get the manifest version.

By default, all tags are considered.  Unless
USE_manifest_version_strip_leading is set, in which case, only those tags
which starts with that will be considered.

As such, this changes the behavior of USE_manifest_version_strip_leading, so
that only tags that matches this will be used (by default).

With the previous behavior, the latest tag would always be used, and if it
matched, would be stripped accordingly.  For non-matching tags, they would
just be used as-is.

The previous behavior can still be achieved by setting
USE_manifest_version_tag_match="*".

With this new behavior, the USE_manifest_version_strip_leading can be safely
used with the initial intent, which is to use it for for relase versioning
using tags with a specific prefix.  The OESTACK_VERSION will always be taken
as relative to the latest release, even if other type of (non-release) tags
have been added since then.

Signed-off-by: Esben Haabendal <esben@haabendal.dk>